Reset Logic Modified / Second Design Laid Out
The first and second designs were modified according to the README specifications, and then the second design's layout was created. The second design's layout unfortunately is encountering a timing issue that was not present in the schematic, and further testing of the circuit is required before a diagnosis can be found.